% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as_csv}
\alias{as_csv}
\title{As csv}
\usage{
as_csv(x, ...)
}
\arguments{
\item{x}{object to parse as csv}

\item{...}{other arguments passed to \link[utils:write.table]{write.csv()}}
}
\value{
object x as parsed csv string
}
\description{
As csv
}
\examples{
as_csv(mtcars)
}
