% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as_datetime_mdy_hms_xm}
\alias{as_datetime_mdy_hms_xm}
\title{As Datetime mm/dd/yyyy hh:mm:ss xm}
\usage{
as_datetime_mdy_hms_xm(x, ...)
}
\arguments{
\item{x}{character representation of datetime}

\item{...}{additional arguments to \link[lubridate:as_date]{lubridate::as_datetime}}
}
\value{
datetime representation of the datetime
}
\description{
As Datetime mm/dd/yyyy hh:mm:ss xm
}
\examples{
as_datetime_mdy_hms_xm("1/1/2001 8:00:32 PM")

}
