% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{box_ftp_download}
\alias{box_ftp_download}
\title{Box FTP Download}
\usage{
box_ftp_download(
  remote,
  local = basename(remote),
  home = Sys.getenv("WUSTL_BOX_HOME"),
  user = Sys.getenv("WUSTL_BOX_USER"),
  pass = Sys.getenv("WUSTL_BOX_PASS"),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{remote}{the path from which the content is to be downloaded.}

\item{local}{file name or path of the local file to write..}

\item{home}{prepended to remote to form the full remote path.}

\item{user}{Box username (i.e., WashU email)}

\item{pass}{unique password for external applications. Created at
\url{https://wustl.app.box.com/account}}

\item{verbose}{emit some progress output}

\item{...}{other arguments passed to \link[curl:handle]{curl::handle_setopt()}}
}
\description{
Box FTP Download
}
\examples{
\dontrun{
box_ftp_download(
  "my_img.png",
  "/dbfs/home/my_user/my_img.png"
)
}
}
