% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{box_write}
\alias{box_write}
\title{Box Write}
\usage{
box_write(x, remote, ...)
}
\arguments{
\item{x}{file contents to write}

\item{remote}{the path to which the content is to be uploaded.}

\item{...}{other arguments passed to \link[curl:handle]{curl::handle_setopt()}}
}
\description{
Box Write
}
\examples{
\dontrun{
box_write("hello world!", "hello world.txt")
}
}
