% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkr.R
\name{collect}
\alias{collect}
\title{collect: Collects all the elements of a SparkDataFrame and coerces...}
\usage{
collect(x, ...)
}
\arguments{
\item{x}{object to collect}

\item{...}{additional arguments

See \verb{SparkR::[collect][SparkR::collect]} for details.}
}
\description{
collect: Collects all the elements of a SparkDataFrame and coerces...
}
