% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.box.R
\name{dbutils.box.read}
\alias{dbutils.box.read}
\title{Box Read}
\usage{
dbutils.box.read(remote, ...)
}
\arguments{
\item{remote}{the path from which the contents are to be read.}

\item{...}{other arguments passed to \link[curl:handle]{curl::handle_setopt()}}
}
\value{
file contents as character string
}
\description{
Box Read
}
\examples{
\dontrun{
dbutils.box.ftps_download("my_sql_file.sql")
}
}
