% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.box.R
\name{dbutils.box.write}
\alias{dbutils.box.write}
\title{Box Write}
\usage{
dbutils.box.write(x, remote, ...)
}
\arguments{
\item{x}{file contents to write}

\item{remote}{the path to which the content is to be uploaded.}

\item{...}{other arguments passed to \link[curl:handle]{curl::handle_setopt()}}
}
\description{
Box Write
}
\examples{
\dontrun{
dbutils.box.write("hello world!", "hello world.txt")
}
}
