% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.cluster.R
\name{dbutils.cluster.persist_app}
\alias{dbutils.cluster.persist_app}
\title{Cluster persist app}
\usage{
dbutils.cluster.persist_app(seconds = 3600)
}
\arguments{
\item{seconds}{seconds to persist}
}
\description{
Cluster persist app
}
