% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.fs.R
\name{dbutils.fs.cp2}
\alias{dbutils.fs.cp2}
\title{Copy content}
\usage{
dbutils.fs.cp2(from, to)
}
\arguments{
\item{from}{from path}

\item{to}{to path}
}
\value{
TRUE if successful
}
\description{
Copy content
}
