% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.fs.R
\name{dbutils.fs.dir}
\alias{dbutils.fs.dir}
\title{List directory}
\usage{
dbutils.fs.dir(x)
}
\arguments{
\item{x}{directory to list}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}} of directory contents
}
\description{
List directory
}
\examples{
\dontrun{
dbutils.fs.dir("~")
}
}
