% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.fs.R
\name{dbutils.fs.exists}
\alias{dbutils.fs.exists}
\title{Path exists}
\usage{
dbutils.fs.exists(x)
}
\arguments{
\item{x}{path to test}
}
\value{
TRUE if path exists
}
\description{
Path exists
}
\examples{
\dontrun{
dbutils.fs.exists("~")
}
}
