% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.fs.R
\name{dbutils.fs.file_store_home}
\alias{dbutils.fs.file_store_home}
\title{FileStore home}
\usage{
dbutils.fs.file_store_home(
  format = c("file", "spark"),
  user = dbutils.credentials.current_user()
)
}
\arguments{
\item{format}{the desired address format (i.e., file: /dbfs/FileStore/... or
spark: dbfs:/FileStore/...)}

\item{user}{name of the user}
}
\value{
path to the FileStore
}
\description{
FileStore home
}
\examples{
dbutils.fs.file_store_home(user = "dborker")
dbutils.fs.file_store_home("spark", "dborker")
}
