% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.fs.R
\name{dbutils.fs.file_store_path}
\alias{dbutils.fs.file_store_path}
\title{FileStore path}
\usage{
dbutils.fs.file_store_path(
  path,
  format = c("file", "spark"),
  user = dbutils.credentials.current_user()
)
}
\arguments{
\item{path}{path from the FileStore home}

\item{format}{the desired address format (i.e., file: /dbfs/FileStore/... or
spark: dbfs:/FileStore/...)}

\item{user}{name of the user}
}
\value{
path to the FileStore file
}
\description{
FileStore path
}
\examples{
dbutils.fs.file_store_path("out.csv", user = "dborker")
dbutils.fs.file_store_path("out.csv", "spark", "dborker")
}
