% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.fs.R
\name{dbutils.fs.file_store_url}
\alias{dbutils.fs.file_store_url}
\title{FileStore URL}
\usage{
dbutils.fs.file_store_url(
  path,
  display_html = TRUE,
  user = dbutils.credentials.current_user(),
  instance = Sys.getenv("DATABRICKS_INSTANCE"),
  org_id = Sys.getenv("DATABRICKS_ORG_ID")
)
}
\arguments{
\item{path}{to the desired file}

\item{display_html}{print as clickable link or just the url as text}

\item{user}{name of the user}

\item{instance}{the databricks instance}

\item{org_id}{the organization id}
}
\description{
FileStore URL
}
\examples{
dbutils.fs.file_store_url(
  "out.csv",
  FALSE,
  "dborker",
  "adb-1234567812345678.12.azuredatabricks.net",
  "1234567890123456"
)
}
