% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.fs.R
\name{dbutils.fs.home}
\alias{dbutils.fs.home}
\title{File system home}
\usage{
dbutils.fs.home(
  type = c("dbfs", "abfs", "file"),
  user = dbutils.credentials.current_user(),
  abfs_group = "data-brokers",
  abfs_host = Sys.getenv("DATABRICKS_ABFSS_HOST")
)
}
\arguments{
\item{type}{which file system}

\item{user}{name of the user}

\item{abfs_group}{group to which the user belongs}

\item{abfs_host}{host name}
}
\value{
path to home directory in desired file system
}
\description{
File system home
}
\examples{
dbutils.fs.home("dbfs", "dborker")
dbutils.fs.home("abfs", "dborker")
dbutils.fs.home("file", "dborker")
}
