% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.home.R
\name{dbutils.home.path}
\alias{dbutils.home.path}
\title{File system home}
\usage{
dbutils.home.path(
  sys = c("file", "dbfs", "filestore", "abfs"),
  fmt = c("file", "spark"),
  user = dbutils.credentials.current_user(),
  abfs_group = "data-brokers",
  abfs_host = Sys.getenv("DATABRICKS_ABFSS_HOST")
)
}
\arguments{
\item{sys}{which file system}

\item{fmt}{how to format the address}

\item{user}{name of the user}

\item{abfs_group}{group to which the user belongs}

\item{abfs_host}{host name}
}
\value{
path to home directory in desired file system
}
\description{
File system home
}
\examples{
dbutils.home.path(user = "dborker")
dbutils.home.path(fmt = "spark", user = "dborker")
dbutils.home.path("dbfs", user = "dborker")
dbutils.home.path("dbfs", "spark", "dborker")
dbutils.home.path("filestore", user = "dborker")
dbutils.home.path("filestore", "spark", "dborker")
dbutils.home.path(
  "abfs",
  user = "dborker",
  abfs_host = "file-share-acmeincprodadls.dfs.core.windows.net"
)
}
