% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.home.R
\name{dbutils.home.persist}
\alias{dbutils.home.persist}
\title{Persist home directory}
\usage{
dbutils.home.persist(
  ephemeral_path = dbutils.home.path("file", "spark"),
  persistent_path = dbutils.home.path("dbfs", "spark")
)
}
\arguments{
\item{ephemeral_path}{path to ephemeral home location}

\item{persistent_path}{path to persistent home location}
}
\value{
TRUE if successful
}
\description{
Persist home directory
}
\examples{
\dontrun{
dbutils.home.persist()
}
}
