% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.home.R
\name{dbutils.home.restore}
\alias{dbutils.home.restore}
\title{Restore home directory}
\usage{
dbutils.home.restore(
  persistent_path = dbutils.home.path("dbfs", "spark"),
  ephemeral_path = dbutils.home.path("file", "spark")
)
}
\arguments{
\item{persistent_path}{path to persistent home location}

\item{ephemeral_path}{path to ephemeral home location}
}
\value{
TRUE if successful
}
\description{
Restore home directory
}
\examples{
\dontrun{
dbutils.home.restore()
}
}
