% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.add_facl_user_sh}
\alias{dbutils.ini.add_facl_user_sh}
\title{Cluster-scoped init script add-facl-\if{html}{\out{<user>}}-\if{html}{\out{<path>}}.sh}
\usage{
dbutils.ini.add_facl_user_sh(
  path,
  user = dbutils.credentials.current_user(),
  perms = "rwx"
)
}
\arguments{
\item{path}{path for which to add user with given permissions}

\item{user}{user to add}

\item{perms}{desired permissions to grant}
}
\description{
Cluster-scoped init script add-facl-\if{html}{\out{<user>}}-\if{html}{\out{<path>}}.sh
}
\examples{
\dontrun{
dbutils.ini.add_facl_user_sh()
}
}
