% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.add_sudo_user_sh}
\alias{dbutils.ini.add_sudo_user_sh}
\title{Cluster-scoped init script add-sudo-\if{html}{\out{<user>}}.sh}
\usage{
dbutils.ini.add_sudo_user_sh(pass, user = dbutils.credentials.current_user())
}
\arguments{
\item{pass}{password}

\item{user}{username}
}
\description{
Cluster-scoped init script add-sudo-\if{html}{\out{<user>}}.sh
}
\examples{
\dontrun{
dbutils.ini.add_sudo_user_sh("rosebud", "dborker")
}
}
