% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.install_databricks_cli_for_root_sh}
\alias{dbutils.ini.install_databricks_cli_for_root_sh}
\title{Cluster-scoped init script install-databricks-cli-for-root.sh}
\usage{
dbutils.ini.install_databricks_cli_for_root_sh()
}
\description{
Cluster-scoped init script install-databricks-cli-for-root.sh
}
\examples{
\dontrun{
dbutils.ini.install_databricks_cli_for_root.sh()
}
}
