% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.install_databricks_cli_for_user_sh}
\alias{dbutils.ini.install_databricks_cli_for_user_sh}
\title{Cluster-scoped init script install-databricks-cli-for-\if{html}{\out{<user>}}.sh}
\usage{
dbutils.ini.install_databricks_cli_for_user_sh(
  user = dbutils.credentials.current_user()
)
}
\arguments{
\item{user}{user account for which to write .databrickscfg}
}
\description{
Cluster-scoped init script install-databricks-cli-for-\if{html}{\out{<user>}}.sh
}
\examples{
\dontrun{
dbutils.ini.install_databricks_cli_for_user_sh()
}
}
