% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.install_odbc_driver_sh}
\alias{dbutils.ini.install_odbc_driver_sh}
\title{Cluster-scoped init script install-odbc-driver.sh}
\usage{
dbutils.ini.install_odbc_driver_sh()
}
\description{
Cluster-scoped init script install-odbc-driver.sh
}
\examples{
\dontrun{
## Generate init script
dbutils.ini.install_odbc_driver_sh()

## R usage of driver
conn <- DBI::dbConnect(
  drv = odbc::odbc(),
  dsn = "Databricks"
)

print(DBI::dbGetQuery(
  conn,
  "SELECT * FROM cleansed.epic_clarity.clarity_orgfilter_pat_enc_hsp LIMIT 2"
))
}
}
