% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.install_openai_sh}
\alias{dbutils.ini.install_openai_sh}
\title{Cluster-scoped init script install-openai.sh}
\usage{
dbutils.ini.install_openai_sh(version = "0.28.1")
}
\arguments{
\item{version}{desired version}
}
\description{
Cluster-scoped init script install-openai.sh
}
\examples{
\dontrun{
dbutils.ini.install_openai_sh()
}
}
