% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.install_rstudio_server_sh}
\alias{dbutils.ini.install_rstudio_server_sh}
\title{Cluster-scoped init script install-rstudio-server.sh}
\usage{
dbutils.ini.install_rstudio_server_sh()
}
\description{
Cluster-scoped init script install-rstudio-server.sh
}
\examples{
\dontrun{
dbutils.ini.install_rstudio_server_sh()
}
}
