% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.restore_directory_sh}
\alias{dbutils.ini.restore_directory_sh}
\title{Cluster-scoped init script for directory restorations}
\usage{
dbutils.ini.restore_directory_sh(persistent_path, ephemeral_path, script_name)
}
\arguments{
\item{persistent_path}{path to the persistent directory}

\item{ephemeral_path}{path to the ephemeral directory}

\item{script_name}{name to use for the script file}
}
\description{
Cluster-scoped init script for directory restorations
}
