% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.restore_home_directory_for_user_sh}
\alias{dbutils.ini.restore_home_directory_for_user_sh}
\title{Cluster-scoped init script restore-home-directory-for-\if{html}{\out{<user>}}.sh}
\usage{
dbutils.ini.restore_home_directory_for_user_sh(
  user = dbutils.credentials.current_user(),
  persistent_path = dbutils.home.path("dbfs", "file", user),
  ephemeral_path = dbutils.home.path("file", "file", user)
)
}
\arguments{
\item{user}{current user}

\item{persistent_path}{path to the persistent home files}

\item{ephemeral_path}{path to the ephemeral home files}
}
\description{
Cluster-scoped init script restore-home-directory-for-\if{html}{\out{<user>}}.sh
}
\examples{
\dontrun{
dbutils.ini.restore_home_directory_for_user_sh()
}
}
