% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.restore_home_directory_sh}
\alias{dbutils.ini.restore_home_directory_sh}
\title{Cluster-scoped init script restore-home-directory.sh}
\usage{
dbutils.ini.restore_home_directory_sh(
  persistent_path = dbutils.home.path("dbfs", "file"),
  ephemeral_path = dbutils.home.path("file", "file")
)
}
\arguments{
\item{persistent_path}{path to the persistent home files}

\item{ephemeral_path}{path to the ephemeral home files}
}
\description{
Cluster-scoped init script restore-home-directory.sh
}
\examples{
\dontrun{
dbutils.ini.restore_home_directory_sh()
}
}
