% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.restore_r_library_for_user_sh}
\alias{dbutils.ini.restore_r_library_for_user_sh}
\title{Cluster-scoped init script restore-r-library-for-\if{html}{\out{<user>}}.sh}
\usage{
dbutils.ini.restore_r_library_for_user_sh(
  user = dbutils.credentials.current_user(),
  persistent_path = dbutils.rlib.path("persistent", "file", user),
  ephemeral_path = dbutils.rlib.path("ephemeral", "file", user)
)
}
\arguments{
\item{user}{current user}

\item{persistent_path}{path to the persistent library}

\item{ephemeral_path}{path to the ephemeral library}
}
\description{
Cluster-scoped init script restore-r-library-for-\if{html}{\out{<user>}}.sh
}
\examples{
\dontrun{
dbutils.ini.restore_r_library_for_user_sh()
}
}
