% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.setup_databricks_cli_sh}
\alias{dbutils.ini.setup_databricks_cli_sh}
\title{Cluster-scoped init script setup-databricks-cli.sh}
\usage{
dbutils.ini.setup_databricks_cli_sh(
  host = Sys.getenv("DATABRICKS_HOST"),
  token = Sys.getenv("DATABRICKS_TOKEN")
)
}
\arguments{
\item{host}{databricks host name}

\item{token}{databricks token}
}
\description{
Cluster-scoped init script setup-databricks-cli.sh
}
\examples{
\dontrun{
dbutils.ini.setup_databricks_cli_sh()
}
}
