% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.setup_rprofile_sh}
\alias{dbutils.ini.setup_rprofile_sh}
\title{Cluster-scoped init script setup-Rprofile.sh}
\usage{
dbutils.ini.setup_rprofile_sh(
  ephemeral_path = dbutils.rlib.path("ephemeral", "file")
)
}
\arguments{
\item{ephemeral_path}{path to the ephemeral library}
}
\description{
Cluster-scoped init script setup-Rprofile.sh
}
\examples{
\dontrun{
dbutils.ini.setup_rprofile_sh()
}
}
