% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.setup_rstudio_server_sh}
\alias{dbutils.ini.setup_rstudio_server_sh}
\title{Cluster-scoped init script setup-rstudio-server.sh}
\usage{
dbutils.ini.setup_rstudio_server_sh(
  studio_deb_url = paste0("https://download2.rstudio.org",
    "/server/bionic/amd64/rstudio-server-2022.12.0-353-amd64.deb")
)
}
\arguments{
\item{studio_deb_url}{url for rstudio deb file}
}
\description{
Cluster-scoped init script setup-rstudio-server.sh
}
\examples{
\dontrun{
dbutils.ini.setup_rstudio_server_sh()
}
}
