% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.write_renviron_for_user_sh}
\alias{dbutils.ini.write_renviron_for_user_sh}
\title{Cluster-scoped init script write-Renviron-for-\if{html}{\out{<user>}}.sh}
\usage{
dbutils.ini.write_renviron_for_user_sh(
  key_values,
  user = dbutils.credentials.current_user()
)
}
\arguments{
\item{key_values}{\code{\link[dplyr:reexports]{dplyr::tibble()}} of \code{key} and \code{value} pairs.}

\item{user}{user for which to write .Renviron}
}
\description{
Cluster-scoped init script write-Renviron-for-\if{html}{\out{<user>}}.sh
}
\examples{
\dontrun{
dbutils.ini.write_renviron_for_user_sh(
  dplyr::tibble(
    key = c("KEY_1", "KEY_2"),
    value = c("VALUE_1", "VALUE_2")
  )
)
}
}
