% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.write_rprofile_for_root_sh}
\alias{dbutils.ini.write_rprofile_for_root_sh}
\title{Cluster-scoped init script write-Rprofile-for-root.sh}
\usage{
dbutils.ini.write_rprofile_for_root_sh(
  x = sprintf(".libPaths(c(\\"\%s\\", .libPaths()))", dbutils.rlib.path("ephemeral",
    "file"))
)
}
\arguments{
\item{x}{.Rprofile contents}
}
\description{
Cluster-scoped init script write-Rprofile-for-root.sh
}
\examples{
\dontrun{
dbutils.ini.write_rprofile_for_root_sh()
}
}
