% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.write_rprofile_for_user_sh}
\alias{dbutils.ini.write_rprofile_for_user_sh}
\title{Cluster-scoped init script write-Rprofile-for-\if{html}{\out{<user>}}.sh}
\usage{
dbutils.ini.write_rprofile_for_user_sh(
  user = dbutils.credentials.current_user(),
  x = sprintf(".libPaths(c(\\"\%s\\", .libPaths()))", dbutils.rlib.path("ephemeral",
    "file", user))
)
}
\arguments{
\item{user}{user for which to write .Rprofile}

\item{x}{.Rprofile contents}
}
\description{
Cluster-scoped init script write-Rprofile-for-\if{html}{\out{<user>}}.sh
}
\examples{
\dontrun{
dbutils.ini.write_rprofile_for_user_sh()
}
}
