% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.write_user_rprofile_sh}
\alias{dbutils.ini.write_user_rprofile_sh}
\title{Cluster-scoped init script write-\if{html}{\out{<user>}}-Rprofile.sh}
\usage{
dbutils.ini.write_user_rprofile_sh(
  x = sprintf(".libPaths(c(\\"\%s\\", .libPaths()))", dbutils.rlib.path("ephemeral",
    "file")),
  user = dbutils.credentials.current_user()
)
}
\arguments{
\item{x}{.Rprofile contents}

\item{user}{user for which to write .Rprofile}
}
\description{
Cluster-scoped init script write-\if{html}{\out{<user>}}-Rprofile.sh
}
\examples{
\dontrun{
dbutils.ini.write_user_rprofile_sh()
}
}
