% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.fs.R
\name{dbutils.rlib.abfs_home}
\alias{dbutils.rlib.abfs_home}
\title{Azure blob file system home}
\usage{
dbutils.rlib.abfs_home(
  group = "data-brokers",
  user = dbutils.credentials.current_user(),
  host = Sys.getenv("DATABRICKS_ABFSS_HOST")
)
}
\arguments{
\item{group}{group to which the user belongs}

\item{user}{name of the user}

\item{host}{host name}
}
\value{
path to home in Azure Blob File System
}
\description{
Azure blob file system home
}
\examples{
dbutils.rlib.abfs_home("data-brokers", "dborker")
}
