% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.fs.R
\name{dbutils.rlib.abfshome}
\alias{dbutils.rlib.abfshome}
\title{Azure blob file system home}
\usage{
dbutils.rlib.abfshome(
  group = "data-brokers",
  user = dbutils.credentials.user(),
  host = "file-share@wusmprodadls.dfs.core.windows.net"
)
}
\arguments{
\item{group}{group to which the user belongs}

\item{user}{name of the user}

\item{host}{host name}
}
\value{
path to home in Azure Blob File System
}
\description{
Azure blob file system home
}
\examples{
dbutils.rlib.abfshome("data-brokers", "dborker")
}
