% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.rlib.R
\name{dbutils.rlib.copy}
\alias{dbutils.rlib.copy}
\title{Copy R library}
\usage{
dbutils.rlib.copy(from, to)
}
\arguments{
\item{from}{from path}

\item{to}{to path}
}
\value{
TRUE if successful
}
\description{
Copy R library
}
