% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.rlib.R
\name{dbutils.rlib.default}
\alias{dbutils.rlib.default}
\title{Set default library}
\usage{
dbutils.rlib.default(path = dbutils.rlib.path("ephemeral"))
}
\arguments{
\item{path}{path to desired library}
}
\description{
Set default library
}
\examples{
\dontrun{
dbutils.rlib.default()
}
}
