% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.rlib.R
\name{dbutils.rlib.details}
\alias{dbutils.rlib.details}
\title{List details of installed packages}
\usage{
dbutils.rlib.details(libpath)
}
\arguments{
\item{libpath}{path(s) to libraries}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}} containing package name, path, and version
}
\description{
List details of installed packages
}
\examples{
dbutils.rlib.details(.libPaths())
}
