% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.rlib.R
\name{dbutils.rlib.ephemeral_path}
\alias{dbutils.rlib.ephemeral_path}
\title{Ephemeral path}
\usage{
dbutils.rlib.ephemeral_path(full = TRUE)
}
\arguments{
\item{full}{pre-pend "file:" to the path}
}
\value{
path to the recommended ephemeral library location
}
\description{
Ephemeral path
}
\examples{
dbutils.rlib.ephemeral_path()
dbutils.rlib.ephemeral_path(full = TRUE)
}
