% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.rlib.R
\name{dbutils.rlib.ephemeralfirst}
\alias{dbutils.rlib.ephemeralfirst}
\title{Ephemeral first}
\usage{
dbutils.rlib.ephemeralfirst(ephemeral_path = dbutils.rlib.ephemeralpath(FALSE))
}
\arguments{
\item{ephemeral_path}{path to ephemeral library location without "file:"}
}
\description{
Ephemeral first
}
\examples{
\dontrun{
dbutils.rlib.ephemeralfirst()
}
}
