% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.rlib.R
\name{dbutils.rlib.ephemeralpath}
\alias{dbutils.rlib.ephemeralpath}
\title{Ephemeral path}
\usage{
dbutils.rlib.ephemeralpath(
  full = TRUE,
  user = dbutils.credentials.currentuser()
)
}
\arguments{
\item{full}{pre-pend "file:" to the path}

\item{user}{username of the current user}
}
\value{
path to the recommended ephemeral library location
}
\description{
Ephemeral path
}
\examples{
dbutils.rlib.ephemeralpath()
dbutils.rlib.ephemeralpath(full = TRUE)
}
