% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.rlib.R
\name{dbutils.rlib.install_spark}
\alias{dbutils.rlib.install_spark}
\title{Install the R Package Spark}
\usage{
dbutils.rlib.install_spark(version = "3.3.1", ...)
}
\arguments{
\item{version}{version to install}

\item{...}{additional arguments passed to \code{\link[pak:pkg_install]{pak::pkg_install()}}}
}
\description{
Install the R Package Spark
}
\examples{
\dontrun{
dbutils.rlib.install_spark()
}
}
