% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.rlib.R
\name{dbutils.rlib.persist}
\alias{dbutils.rlib.persist}
\title{Persist R library}
\usage{
dbutils.rlib.persist(
  ephemeral_path = dbutils.rlib.path("ephemeral"),
  persistent_path = dbutils.rlib.path("persistent")
)
}
\arguments{
\item{ephemeral_path}{path to ephemeral library location}

\item{persistent_path}{path to persistent library location}
}
\value{
TRUE if successful
}
\description{
Persist R library
}
\examples{
\dontrun{
dbutils.rlib.persist()
}
}
