% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.rlib.R
\name{dbutils.rlib.persistentpath}
\alias{dbutils.rlib.persistentpath}
\title{Persistent path}
\usage{
dbutils.rlib.persistentpath(
  group = "data-brokers",
  user = dbutils.credentials.currentuser(),
  host = "file-share@wusmprodadls.dfs.core.windows.net"
)
}
\arguments{
\item{group}{group to which the user belongs}

\item{user}{name of the user}

\item{host}{host name}
}
\value{
path to the recommended persistent path in Azure Blob File System
}
\description{
Persistent path
}
\examples{
dbutils.rlib.persistentpath("data-brokers", "dborker")
}
