% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.rlib.R
\name{dbutils.rlib.restore}
\alias{dbutils.rlib.restore}
\title{Restore R library}
\usage{
dbutils.rlib.restore(
  persistent_path = dbutils.rlib.persistentpath(),
  ephemeral_path = dbutils.rlib.ephemeralpath()
)
}
\arguments{
\item{persistent_path}{path to persistent library location}

\item{ephemeral_path}{path to ephemeral library location}
}
\value{
TRUE if successful
}
\description{
Restore R library
}
\examples{
\dontrun{
dbutils.rlib.restore()
}
}
