% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.secrets.R
\name{dbutils.secrets.list_scopes}
\alias{dbutils.secrets.list_scopes}
\title{List Scopes}
\usage{
dbutils.secrets.list_scopes(
  token = Sys.getenv("DATABRICKS_PAT"),
  instance = Sys.getenv("DATABRICKS_INSTANCE")
)
}
\arguments{
\item{token}{A databricks API personal access token.}

\item{instance}{A databricks instance (e.g.,
dbc-a1b2345c-d6e7.cloud.databricks.com)}
}
\value{
\code{\link[httr2:response]{httr2::response()}}
}
\description{
List Scopes
}
\examples{
\dontrun{
dbutils.secrets.list_scopes()
}
}
