% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.secrets.R
\name{dbutils.secrets.load}
\alias{dbutils.secrets.load}
\title{Load secrets}
\usage{
dbutils.secrets.load(scope)
}
\arguments{
\item{scope}{name of the desired scope/key vault}
}
\description{
Load secrets
}
\examples{
\dontrun{
dbutils.secrets.load("wusm-prod-biostats-kv")
}
}
