% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.secrets.R
\name{dbutils.secrets.load.wusm_prod_databrokers_kv}
\alias{dbutils.secrets.load.wusm_prod_databrokers_kv}
\title{Load secrets from the Databrokers Production Key Vault}
\usage{
dbutils.secrets.load.wusm_prod_databrokers_kv(pattern = ".+")
}
\arguments{
\item{pattern}{character string containing a regular expression to select
which vault secrets to load}
}
\description{
Load secrets from the Databrokers Production Key Vault
}
\examples{
\dontrun{
dbutils.secrets.load.wusm_prod_databrokers_kv()
}

}
