% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.secrets.R
\name{dbutils.secrets.tbl_df}
\alias{dbutils.secrets.tbl_df}
\title{Retrieve secrets as a tibble}
\usage{
dbutils.secrets.tbl_df(scope, pattern)
}
\arguments{
\item{scope}{name of the desired scope/key vault}

\item{pattern}{character string containing a regular expression to select
which vault secrets to include}
}
\description{
Retrieve secrets as a tibble
}
\examples{
\dontrun{
dbutils.secrets.tbl_df("wusm-prod-biostats-kv", ".+")
}
}
