% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.secrets.R
\name{dbutils.secrets.tbl_df.wusm_prod_databrokers_kv}
\alias{dbutils.secrets.tbl_df.wusm_prod_databrokers_kv}
\title{Retrieve secrets as a tibble from the Databrokers Production Key Vault}
\usage{
dbutils.secrets.tbl_df.wusm_prod_databrokers_kv(select = ".+", ...)
}
\arguments{
\item{select}{character string containing a regular expression to be matched
when selecting key vault secrets}

\item{...}{further arguments passed to or from other methods}
}
\description{
Retrieve secrets as a tibble from the Databrokers Production Key Vault
}
\examples{
\dontrun{
dbutils.secrets.tbl_df.wusm_prod_databrokers_kv()
}

}
