% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{print.init_script}
\alias{print.init_script}
\title{Print an Init-Script}
\usage{
\method{print}{init_script}(x)
}
\arguments{
\item{x}{object of class init_script}
}
\description{
Print an Init-Script
}
