% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rstudio_server_url}
\alias{rstudio_server_url}
\title{RStudio Server URL}
\usage{
rstudio_server_url(
  display_html = TRUE,
  host = Sys.getenv("DATABRICKS_HOST"),
  org_id = Sys.getenv("DATABRICKS_ORG_ID"),
  cluster_id = Sys.getenv("DATABRICKS_CLUSTER_ID"),
  port = 8787
)
}
\arguments{
\item{display_html}{print as clickable link or just the url as text}

\item{host}{the databricks host}

\item{org_id}{the organization id}

\item{cluster_id}{the cluster id}

\item{port}{port for the server}
}
\description{
RStudio Server URL
}
\examples{
rstudio_server_url(
  display_html = FALSE,
  host = "adb-1234567812345678.12.azuredatabricks.net",
  org_id = "1234567890123456",
  cluster_id = "1234-123456-1234abcd"
)
}
