% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.setup.R
\name{write.setup.r}
\alias{write.setup.r}
\title{Write cluster-scoped init script setup R}
\usage{
write.setup.r(
  persistent_path = dbutils.rlib.path("persistent"),
  ephemeral_path = dbutils.rlib.path("ephemeral")
)
}
\arguments{
\item{persistent_path}{path to the persistent library}

\item{ephemeral_path}{path to the ephemeral library}
}
\value{
TRUE if successful
}
\description{
Write cluster-scoped init script setup R
}
\examples{
\dontrun{
write.setup.r()
}
}
