% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.setup.R
\name{write.setup.renviron}
\alias{write.setup.renviron}
\title{Write cluster-scoped init script to generate .Renviron}
\usage{
write.setup.renviron(key_values)
}
\arguments{
\item{key_values}{\code{\link[dplyr:reexports]{dplyr::tibble()}} of \code{key} and \code{value} pairs.}
}
\value{
TRUE if successful
}
\description{
Write cluster-scoped init script to generate .Renviron
}
\examples{
\dontrun{
write.setup.renviron(
  dplyr::tibble(
    key = c("KEY_1", "KEY_2"),
    value = c("VALUE_1", "VALUE_2")
  )
)
}
}
