% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.setup.R
\name{write.setup.rprofile}
\alias{write.setup.rprofile}
\title{Write cluster-scoped init script to generate .Rprofile}
\usage{
write.setup.rprofile(ephemeral_path = dbutils.rlib.path("ephemeral"))
}
\arguments{
\item{ephemeral_path}{path to the ephemeral library}
}
\value{
TRUE if successful
}
\description{
Write cluster-scoped init script to generate .Rprofile
}
\examples{
\dontrun{
write.setup.rprofile()
}
}
