% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wubik_install}
\alias{wubik_install}
\title{Install {wubik}}
\usage{
wubik_install(pkg = "the-mad-statter/wubik", ...)
}
\arguments{
\item{pkg}{wubik github reference}

\item{...}{additional arugments passed to \code{\link[pak:pkg_install]{pak::pkg_install()}}}
}
\value{
(Invisibly) A data frame with information about the installed
package(s).
}
\description{
Install {wubik}
}
\examples{
\dontrun{
wubik_install()
}
}
