% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{box_read}
\alias{box_read}
\title{Box Read}
\usage{
box_read(remote, ...)
}
\arguments{
\item{remote}{the path from which the contents are to be read.}

\item{...}{other arguments passed to \link[curl:handle]{curl::handle_setopt()}}
}
\value{
file contents as character string
}
\description{
Box Read
}
\examples{
\dontrun{
box_read("my_sql_file.sql")
}
}
