% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.credentials.R
\name{dbutils.credentials.current_user}
\alias{dbutils.credentials.current_user}
\title{Current user}
\usage{
dbutils.credentials.current_user()
}
\value{
username of the current user
}
\description{
Current user
}
\examples{
\dontrun{
dbutils.credentials.current_user()
}
}
