% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.credentials.R
\name{dbutils.credentials.current_user}
\alias{dbutils.credentials.current_user}
\title{Current user}
\usage{
dbutils.credentials.current_user(x)
}
\arguments{
\item{x}{current user if known}
}
\value{
username of the current user
}
\description{
Current user
}
\examples{
dbutils.credentials.current_user("dborker@wustl.edu")
\dontrun{
#' dbutils.credentials.current_user()
}
}
