% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.fs.R
\name{dbutils.fs.home}
\alias{dbutils.fs.home}
\title{File system home}
\usage{
dbutils.fs.home(
  system = c("file", "dbfs", "abfs"),
  format = c("file", "spark"),
  user = dbutils.credentials.current_user(),
  abfs_group = "data-brokers",
  abfs_host = Sys.getenv("DATABRICKS_ABFSS_HOST")
)
}
\arguments{
\item{system}{which file system}

\item{format}{how to format the address}

\item{user}{name of the user}

\item{abfs_group}{group to which the user belongs}

\item{abfs_host}{host name}
}
\value{
path to home directory in desired file system
}
\description{
File system home
}
\examples{
dbutils.fs.home(user = "dborker")
dbutils.fs.home(format = "spark", user = "dborker")
dbutils.fs.home("dbfs", user = "dborker")
dbutils.fs.home("dbfs", "spark", "dborker")
dbutils.fs.home(
  "abfs",
  user = "dborker",
  abfs_host = "file-share-acmeincprodadls.dfs.core.windows.net"
)
}
