% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.fs.R
\name{dbutils.fs.home}
\alias{dbutils.fs.home}
\title{File system home}
\usage{
dbutils.fs.home(
  type = c("file", "dbfs", "abfs"),
  canonical = FALSE,
  user = dbutils.credentials.current_user(),
  abfs_group = "data-brokers",
  abfs_host = Sys.getenv("DATABRICKS_ABFSS_HOST")
)
}
\arguments{
\item{type}{which file system}

\item{canonical}{whether to prepend with file system (e.g., file: or dbfs:)}

\item{user}{name of the user}

\item{abfs_group}{group to which the user belongs}

\item{abfs_host}{host name}
}
\value{
path to home directory in desired file system
}
\description{
File system home
}
\examples{
dbutils.fs.home(user = "dborker")
dbutils.fs.home(canonical = TRUE, user = "dborker")
dbutils.fs.home("dbfs", user = "dborker")
dbutils.fs.home("dbfs", TRUE, "dborker")
dbutils.fs.home(
  "abfs",
  user = "dborker",
  abfs_host = "file-share-acmeincprodadls.dfs.core.windows.net"
)
}
