% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.install_odbc_driver_sh}
\alias{dbutils.ini.install_odbc_driver_sh}
\title{Cluster-scoped init script install-odbc-driver.sh}
\usage{
dbutils.ini.install_odbc_driver_sh()
}
\description{
Cluster-scoped init script install-odbc-driver.sh
}
\note{
Because this script will pull values from Spark environment variables:
DATABRICKS_HOST, DATABRICKS_HTTP_PATH, and DATABRICKS_TOKEN, ensure these
environment variables are present during cluster imitation. To do so it is
recommended that these values be stored in an Azure Key Vault and set in the
cluster configuration UI: Spark -> Environment variables. For example:
\code{DATABRICKS_TOKEN={{secrets/wusm-prod-biostats-kv/DATABRICKS-TOKEN}}}.
Connection can be tested in a terminal with \verb{isql -v Databricks}.
}
\examples{
\dontrun{
## Generate init script
dbutils.ini.install_odbc_driver_sh()

## R usage of driver
conn <- DBI::dbConnect(
  drv = odbc::odbc(),
  dsn = "Databricks"
)

print(DBI::dbGetQuery(
  conn,
  "SELECT * FROM cleansed.epic_clarity.clarity_orgfilter_pat_enc_hsp LIMIT 2"
))
}

}
