% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.write}
\alias{dbutils.ini.write}
\title{Write cluster-scoped init script}
\usage{
dbutils.ini.write(x, name = attr(x, "name"))
}
\arguments{
\item{x}{script contents to write}

\item{name}{name of the script}
}
\value{
TRUE if successful
}
\description{
Write cluster-scoped init script
}
\examples{
\dontrun{
dbutils.ini.write(dbutils.ini.setup_r_sh())
dbutils.ini.write(dbutils.ini.setup_rprofile_sh())
dbutils.ini.write(dbutils.ini.setup_renviron_sh())
dbutils.ini.write(dbutils.ini.setup_databricks_cli_sh())
dbutils.ini.write(dbutils.ini.setup_rstudio_server_sh())
dbutils.ini.write(dbutils.ini.setup_user_sh())
}
}
