% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.ini.R
\name{dbutils.ini.write}
\alias{dbutils.ini.write}
\title{Write cluster-scoped init script}
\usage{
dbutils.ini.write(
  x,
  name = attr(x, "name"),
  user = dbutils.credentials.current_user(),
  path = sprintf("dbfs:/databricks/scripts/\%s/\%s", user, name)
)
}
\arguments{
\item{x}{script contents to write}

\item{name}{name of the script}

\item{user}{current user}

\item{path}{path to write script}
}
\value{
TRUE if successful
}
\description{
Write cluster-scoped init script
}
\examples{
\dontrun{
## Restore R Library for Current User
dbutils.ini.write(dbutils.ini.restore_r_library_for_user_sh())

## Write .Rprofile(s)
### for Current User
dbutils.ini.write(dbutils.ini.write_rprofile_for_user_sh())
### for Root
dbutils.ini.write(dbutils.ini.write_rprofile_for_root_sh())

## Write .Renviron
kvps <- tibble(key = "key_name", value = "key_value")
### for Current User
dbutils.ini.write(dbutils.ini.write_renviron_for_user_sh(kvps))
### for Root
dbutils.ini.write(dbutils.ini.write_renviron_for_root_sh(kvps))

## Databricks CLI
### for Current User
dbutils.ini.write(dbutils.ini.install_databricks_cli_for_user_sh())
### for Root
dbutils.ini.write(dbutils.ini.install_databricks_cli_for_root_sh())

## Install RStudio Server
dbutils.ini.write(dbutils.ini.install_rstudio_server_sh())

## Add Current User as Super User with Password
dbutils.ini.write(dbutils.ini.add_sudo_user_sh("rosebud"))

## Install ODBC Driver
dbutils.ini.write(dbutils.ini.install_odbc_driver_sh())

## Edit (Home) Directory Permissions for Current User
dbutils.ini.write(dbutils.ini.add_facl_user_to_path_sh())
}
}
