% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbutils.secrets.R
\name{dbutils.secrets.load.wusm_prod_biostats_kv}
\alias{dbutils.secrets.load.wusm_prod_biostats_kv}
\title{Load secrets from the Biostats Production Key Vault}
\usage{
dbutils.secrets.load.wusm_prod_biostats_kv(select = ".+", ...)
}
\arguments{
\item{select}{character string containing a regular expression to be matched
when selecting key vault secrets}

\item{...}{further arguments passed to or from other methods}
}
\description{
Load secrets from the Biostats Production Key Vault
}
\examples{
\dontrun{
dbutils.secrets.load.wusm_prod_biostats_kv()
}

}
