% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.setup.R
\name{write.setup.renviron}
\alias{write.setup.renviron}
\title{Write cluster-scoped init script to generate .Renviron}
\usage{
write.setup.renviron()
}
\value{
TRUE if successful
}
\description{
Write cluster-scoped init script to generate .Renviron
}
\examples{
\dontrun{
write.setup.renviron()
}
}
